[FIX] models.py: onchange creating m2m => no NewId
Since 9f2dbf7 if we created a record in a many2many relationship of an
onchange, we would also possibly erreneously return the NewId.

The logic of convert_to_onchange of _RelationalMulti doesn't send back
the ID field (if the ID is needed, it is part of the command, not the
update/create dictionary values of a command) and this commit does the
same for the code changed in 9f2dbf7.

This commit also adds a test for this fix, and a test for the regression
corrected with 6695b86291 reported in #25969

fix for https://github.com/odoo/odoo/commit/9f2dbf72a#commitcomment-29837226
closes #25984