[REF] point of sale: revert wrong configuration change
This commit reverts a configuration update recently done. Tax display
configuration was moved from pos config to general company settings.
This was a mistake and should not stay as it is. Even if saas 11.4 could
be considered as stable this change is necessary and does not impact
anybody as this branch is not released.

Partially revert 8dce53dcc32a703e1bc7a0f595b6c1dd99df6c3c .