Revert "[FIX] models: make `onchange` return a smaller diff"
This reverts commit 9f2dbf72abea18f6d1c23c0477d5214ff305b177 which
already has been fixed multiple times [0].

However this change does not correctly handle successive onchange calls on
ony2many fields which alter existing rows.

i.e., in saas~11.4, an onchange has been added on move lines [1] in
order to create and/or update lines containing taxes amounts.
However, with diff-only onchange, a second call to onchange will discard
changes made by a first one as nothing change between the 2 calls.

Scenario: On an existing account.move, add a new credit line with a tax set.
The first onchange will update the tax line with new amount. Add a
second line (with debit amount balanced). The onchange will not change
the tax line, resulting as it value being reset to initial value. As a
result, the move is not balanced and cannot be saved.

[0] 6695b862915ec9e0bd3ee215b9471afc9feb9532, f415a96588db601c866238e91007198c48afbe70, f726f7843054d0c453504e3d2cbb858941c647c9
[1] bf1fb9335470a3d5edf73576df6c453462f97e9d