[FIX] mrp: see full name of 'report' option in BoM 'Structure & Cost' report
Before this commit, when we opened the report "Structure & Cost" of a Bill of
Materials of a product without any variant, the dropdown option selection
'report' had the current selection name cut. For instance, the selection
"BoM Structure & Cost" was displayed as "BoM Structure".

This was due to each option having a width determined by the bootstrap
classname `col-md-4`, independently on the number of options that are
available.

There are either 2 or 3 options at the top of this report, depending on whether
the product of the BoM has variants or not. When there are 3 options, there is
not problem. However, with 2 options (i.e., without variants), each option
available horizontal space is reduced, so that the 'report' option is shrunk,
which cuts the text selection as a result.

This commit fixes this problem when there are 2 options (= "no variant"),
so that the 'report' option uses more horizontal space.

Task-ID 1864490
Closes #25643