[IMP] product: Choose volume measure on products
Until now, the volume measure displayed on the product form was only in
m³. We now allow to choose between m³ and ft³.

So, we add volume selection field (m³/ft³) on general settings and it will
be display on product and product variants form view as configuration.

We also Remove tooltip from weight and volume field.

And display weight (kg or lb(s)) in product and product variants form view correctly,
because currently weight didn't display on product form view before save the product
but after this commit weight will be display.

This commit is related to task ID 1869002.

closes odoo/odoo#26542