[IMP] mass_mailing, crm, sale: improve demo data related to UTM
Purpose is to be able to have some basic demo data for crm and sales linked
to UTMs. Those will be used when introducing crm and sale bridge module with
mass mailing in order to display some documents linked by UTMs.

This commit is linked to task ID 1872198 and PR #26188.